+++ CHANGELOG.md
@@ -7,7 +7,10 @@
 - TT-1199 - Fixing accessibility issues
 - TT-2025 - Update success page content
 - No ticket - Removed old feature flagged items
+- TT-1608-member-request-admin-2
+- TT-1608 - Member To Request Admin
 - No ticket - upgrade django
+
 
 ### Fixed bugs:
 - TT-1817 - Use correct input types for number fields